Initial release
Entire code is contained in a single file. Designed to be compiled and executed as a command line program. Any changes in file output is up to the user. The output files are geometry information in the OFF format. The 'DisplaySphere' program can be used to display the OFF files as well as other available OFF viewing software.